ALL Favorite by id utilisateur

+++ src/main/java/com/movieseries/movieseries/dto/UtilisateurDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UtilisateurDTO {
-    private int id_Utilisateur;
+    private Long id_Utilisateur;
     private String email;
     private String nom;
     private String password;
